+++ src/common.rs
@@ -38,15 +38,15 @@ pub(crate) use crate::{
 pub(crate) use crate::{
   alias::Alias, alias_resolver::AliasResolver, assignment_evaluator::AssignmentEvaluator,
   assignment_resolver::AssignmentResolver, color::Color, compilation_error::CompilationError,
-  compilation_error_kind::CompilationErrorKind, configuration::Configuration,
-  expression::Expression, fragment::Fragment, function::Function,
-  function_context::FunctionContext, functions::Functions, interrupt_guard::InterruptGuard,
-  interrupt_handler::InterruptHandler, justfile::Justfile, lexer::Lexer, output_error::OutputError,
-  parameter::Parameter, parser::Parser, platform::Platform, position::Position, recipe::Recipe,
-  recipe_context::RecipeContext, recipe_resolver::RecipeResolver, runtime_error::RuntimeError,
-  search_error::SearchError, shebang::Shebang, state::State, string_literal::StringLiteral,
-  token::Token, token_kind::TokenKind, use_color::UseColor, variables::Variables,
-  verbosity::Verbosity, warning::Warning,
+  compilation_error_kind::CompilationErrorKind, config::Config, expression::Expression,
+  fragment::Fragment, function::Function, function_context::FunctionContext, functions::Functions,
+  interrupt_guard::InterruptGuard, interrupt_handler::InterruptHandler, justfile::Justfile,
+  lexer::Lexer, output_error::OutputError, parameter::Parameter, parser::Parser,
+  platform::Platform, position::Position, recipe::Recipe, recipe_context::RecipeContext,
+  recipe_resolver::RecipeResolver, runtime_error::RuntimeError, search_error::SearchError,
+  shebang::Shebang, state::State, string_literal::StringLiteral, token::Token,
+  token_kind::TokenKind, use_color::UseColor, variables::Variables, verbosity::Verbosity,
+  warning::Warning,
 };
 
 pub(crate) type CompilationResult<'a, T> = Result<T, CompilationError<'a>>;
