+++ src/recipe_context.rs
@@ -2,6 +2,6 @@ use crate::common::*;
 
 pub(crate) struct RecipeContext<'a> {
   pub(crate) invocation_directory: &'a Result<PathBuf, String>,
-  pub(crate) configuration: &'a Configuration<'a>,
+  pub(crate) config: &'a Config<'a>,
   pub(crate) scope: BTreeMap<&'a str, String>,
 }
